Fix height of schedule view event, center event icon  #16
